refactor: refactor types folder
