--- src/db/types/auth.py
@@ -1,6 +1,0 @@
-from utils.schemas import BaseModel
-
-
-class JWTPair(BaseModel):
-    access_token: str
-    refresh_token: str
