+++ src/db/types/types.py
@@ -1,7 +1,19 @@
-from pydantic import Field
+from typing import Literal
 
+from pydantic import Field
 from utils.schemas import BaseModel
 
+
+class JWTPair(BaseModel):
+    access_token: str
+    refresh_token: str
+
+EntityTargetType = Literal["user", "group", "contest"]
+
+class Entity(BaseModel):
+    id: str = Field(alias='_id')
+    target_type: EntityTargetType
+    target_id: int
 
 class _UserBase(BaseModel):
     domain: str | None = None
