+++ src/routers/test.py
@@ -6,9 +6,8 @@ from loguru import logger
 import utils.auth
 from config import config
 from utils.response import ErrorCodes, ErrorResponse, SuccessfulResponse
-from db import client, methods, types
-from db.types.requests import RQ
-from db.types.responses import RS
+from db import client, methods
+from db.types import types, RS, RQ
 
 
 router = APIRouter()
@@ -27,7 +26,7 @@ async def signup(request: RQ.test.signup):
     hashed_password = utils.auth.hash_password(request.password)
 
     inserted_user_id = methods.users.insert_user_with_id(
-        types.user.UserWithoutID(
+        types.UserWithoutID(
             email=request.email,
             hashed_password=hashed_password,
             first_name=request.first_name
